+++ unionfind/src/main/java/regular_union_find/ICorrelator.java
@@ -1,4 +1,4 @@
-package algorithm;
+package regular_union_find;
 
 import java.util.List;
 import java.util.Set;
