+++ unionfind/src/main/java/regular_union_find/UnionFind.java
@@ -1,4 +1,4 @@
-package algorithm;
+package regular_union_find;
 
 import java.util.HashMap;
 import java.util.HashSet;
